+++ src/hardware/joystick.cpp
@@ -36,7 +36,7 @@
 struct JoyStick {
 	bool enabled;
 	float xpos,ypos;
-	double xtick,ytick;
+	float xtick,ytick;
 	Bitu xcount,ycount;
 	bool button[2];
 };
@@ -89,7 +89,7 @@ static Bitu read_p201(Bitu port,Bitu iolen) {
 
 static Bitu read_p201_timed(Bitu port,Bitu iolen) {
 	Bit8u ret=0xff;
-	double currentTick = PIC_FullIndex();
+	float currentTick = PIC_FullIndex();
 	if( stick[0].enabled ){
 		if( stick[0].xtick < currentTick ) ret &=~1;
 		if( stick[0].ytick < currentTick ) ret &=~2;
@@ -129,18 +129,18 @@ static void write_p201_timed(Bitu port,Bitu val,Bitu iolen) {
 	// Axes take time = 24.2 microseconds + ( 0.011 microsecons/ohm * resistance )
 	// to reset to 0
 	// Precalculate the time at which each axis hits 0 here
-	double currentTick = PIC_FullIndex();
+	float currentTick = PIC_FullIndex();
 	if (stick[0].enabled) {
 		stick[0].xtick = currentTick + 1000.0*( JOY_S_CONSTANT + S_PER_OHM *
-	                         (double)(((stick[0].xpos+1.0)* OHMS)) );
+	                         (float)(((stick[0].xpos+1.0)* OHMS)) );
 		stick[0].ytick = currentTick + 1000.0*( JOY_S_CONSTANT + S_PER_OHM *
-		                 (double)(((stick[0].ypos+1.0)* OHMS)) );
+		                 (float)(((stick[0].ypos+1.0)* OHMS)) );
 	}
 	if (stick[1].enabled) {
 		stick[1].xtick = currentTick + 1000.0*( JOY_S_CONSTANT + S_PER_OHM *
-		                 (double)((swap34? stick[1].ypos : stick[1].xpos)+1.0) * OHMS);
+		                 (float)((swap34? stick[1].ypos : stick[1].xpos)+1.0) * OHMS);
 		stick[1].ytick = currentTick + 1000.0*( JOY_S_CONSTANT + S_PER_OHM *
-		                 (double)((swap34? stick[1].xpos : stick[1].ypos)+1.0) * OHMS);
+		                 (float)((swap34? stick[1].xpos : stick[1].ypos)+1.0) * OHMS);
 	}
 }
